refactor: 스타일드 컴포넌트 파일명 변경( style -> styles )
